requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0

psycopg[binary]==3.2.1
redis==5.0.8

pytest==8.3.2
httpx==0.27.2

pypdf==4.0.2
numpy==1.26.4

faiss-cpu==1.8.0.post1
numpy==1.26.4
sentence-transformers==2.7.0
torch==2.2.2
